requirements.txt
astroid==2.3.3
azure-common==1.1.25
azure-functions==1.2.0
azure-mgmt-redis==6.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
idna==2.9
isodate==0.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msrest==0.6.11
msrestazure==0.6.2
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.1
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
